Fix update and delete user endpoints to also update and delete identity tables

+++ PetitesVictoires/PetitesVictoires.UseCases/Users/Delete/DeleteUserHandler.cs
@@ -1,18 +1,27 @@
 using Ardalis.Result;
 using Ardalis.SharedKernel;
 using Mediator;
+using PetitesVictoires.Core.Interfaces;
 using PetitesVictoires.Core.UserAggregate;
 
 namespace PetitesVictoires.UseCases.Users.Delete;
 
-public class DeleteUserHandler(IRepository<User> repository) : ICommandHandler<DeleteUserCommand, Result>
+public class DeleteUserHandler(IRepository<User> repository, IIdentityService identityService, IUnitOfWork unitOfWork)
+    : ICommandHandler<DeleteUserCommand, Result>
 {
     public async ValueTask<Result> Handle(DeleteUserCommand command, CancellationToken cancellationToken)
     {
         var userToDelete = await repository.GetByIdAsync(command.UserId, cancellationToken);
         if (userToDelete is null) return Result.NotFound();
 
+        await using var transaction = await unitOfWork.BeginTransactionAsync(cancellationToken);
+
+        var identityResult = await identityService.DeleteUserAsync(userToDelete.Id, cancellationToken);
+        if (!identityResult.IsSuccess) return identityResult;
+
         await repository.DeleteAsync(userToDelete, cancellationToken);
+
+        await transaction.CommitAsync(cancellationToken);
         return Result.Success();
     }
 }
